dlang: small tweaks to get nm functions to compile in and x86_64 system.
Use of size_t to catch array lengths is the portable wat to keep uint/ulong
data on 32/64-bit machines.

+++ dlang/nm/gaussquad.d
@@ -40,7 +40,7 @@ double apply_gauss_rule(alias f)(double a, double b, double[] xs, double[] ws)
 {
     double xmid = 0.5 * (b + a);
     double xrange2 = 0.5 * (b - a);
-    uint N = xs.length;
+    size_t N = xs.length;
     double result = 0.0;
     foreach(i; 0 .. N) {
 	double x = xmid + xs[i] * xrange2; // transform from range -1.0 +1.0
